requirements.txt
flask>=2.0.1
pandas>=2.0.0
openpyxl>=3.0.9
tabula-py>=2.9.0
PyPDF2>=3.0.0
python-evtx>=0.7.4 